Refactored logic to use integer instead of big integer.

+++ src/main/java/org/elece/sql/parser/expression/internal/SqlNumberValue.java
@@ -2,13 +2,13 @@ package org.elece.sql.parser.expression.internal;
 
 import java.math.BigInteger;
 
-public class SqlNumberValue extends SqlValue<BigInteger> {
-    public SqlNumberValue(BigInteger value) {
+public class SqlNumberValue extends SqlValue<Integer> {
+    public SqlNumberValue(Integer value) {
         super(value);
     }
 
     @Override
-    protected Integer compare(SqlValue<BigInteger> target) {
+    protected Integer compare(SqlValue<Integer> target) {
         return getValue().compareTo(target.getValue());
     }
 }
